код юли не рабочий

+++ main/main.cpp
@@ -45,7 +45,7 @@ int main() {
     // Вычисление определителя
     try {
         auto start = high_resolution_clock::now();
-        double determinant = mat.determinant();
+        int determinant = mat.determinant();
         auto end = high_resolution_clock::now();
         duration<double> elapsed = end - start;
         cout << "Определитель матрицы: " << determinant << endl;
